fix: rename task model (contract) file same as other
