+++ packages/contracts/src/goals.model.ts
@@ -1,7 +1,7 @@
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
 import { IEmployee, IEmployeeFindInput } from './employee.model';
 import { IOrganizationProject } from './organization-projects.model';
-import { ITask } from './task-entity.model';
+import { ITask } from './task.model';
 import { IOrganizationTeam } from './organization-team-model';
 import { IKPI } from './goal-settings.model';
