+++ packages/contracts/src/organization-projects.model.ts
@@ -7,7 +7,7 @@ import {
 } from './organization.model';
 import { IBaseEntityWithMembers } from './entity-with-members.model';
 import { ITag } from './tag-entity.model';
-import { ITask } from './task-entity.model';
+import { ITask } from './task.model';
 import { IOrganizationSprint } from './organization-sprint.model';
 import { IPayment } from './payment.model';
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
