+++ packages/contracts/src/organization-team-model.ts
@@ -1,7 +1,7 @@
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
 import { IOrganizationTeamEmployee } from './organization-team-employee-model';
 import { ITag } from './tag-entity.model';
-import { ITask } from './task-entity.model';
+import { ITask } from './task.model';
 import { ITimerStatusInput } from './timesheet.model';
 
 export interface IOrganizationTeam extends IBasePerTenantAndOrganizationEntityModel {
